+++ doma-core/src/main/java/org/seasar/doma/internal/jdbc/scalar/OptionalDoubleScalarSuppliers.java
@@ -1,0 +1,11 @@
+package org.seasar.doma.internal.jdbc.scalar;
+
+import java.util.OptionalDouble;
+import java.util.function.Supplier;
+
+public final class OptionalDoubleScalarSuppliers {
+
+  public static Supplier<Scalar<Double, OptionalDouble>> of() {
+    return OptionalDoubleScalar::new;
+  }
+}