+++ doma-core/src/main/java/org/seasar/doma/internal/jdbc/scalar/OptionalIntScalarSuppliers.java
@@ -1,0 +1,11 @@
+package org.seasar.doma.internal.jdbc.scalar;
+
+import java.util.OptionalInt;
+import java.util.function.Supplier;
+
+public final class OptionalIntScalarSuppliers {
+
+  public static Supplier<Scalar<Integer, OptionalInt>> of() {
+    return OptionalIntScalar::new;
+  }
+}