+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/context/SubSelectContext.java
@@ -1,11 +1,28 @@
 package org.seasar.doma.jdbc.criteria.context;
 
 import java.util.Objects;
+import org.seasar.doma.jdbc.criteria.metamodel.PropertyMetamodel;
 
-public class SubSelectContext<RESULT> {
-  public final SelectContext context;
+public interface SubSelectContext<RESULT> {
 
-  public SubSelectContext(SelectContext context) {
-    this.context = Objects.requireNonNull(context);
+  SelectContext get();
+
+  class Single<PROPERTY> implements SubSelectContext<PropertyMetamodel<PROPERTY>> {
+    private final SelectContext context;
+    private final PropertyMetamodel<PROPERTY> propertyMetamodel;
+
+    public Single(SelectContext context, PropertyMetamodel<PROPERTY> propertyMetamodel) {
+      this.context = Objects.requireNonNull(context);
+      this.propertyMetamodel = Objects.requireNonNull(propertyMetamodel);
+    }
+
+    @Override
+    public SelectContext get() {
+      return context;
+    }
+
+    public PropertyMetamodel<PROPERTY> getPropertyMetamodel() {
+      return propertyMetamodel;
+    }
   }
 }