+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/declaration/SubSelectFromDeclaration.java
@@ -12,7 +12,7 @@ import org.seasar.doma.jdbc.criteria.metamodel.EntityMetamodel;
 import org.seasar.doma.jdbc.criteria.metamodel.PropertyMetamodel;
 import org.seasar.doma.jdbc.criteria.tuple.Tuple2;
 
-public class SubSelectFromDeclaration {
+public class SubSelectFromDeclaration<ENTITY> implements SubSelectContext<ENTITY> {
 
   private final SelectFromDeclaration declaration;
 
@@ -22,54 +22,69 @@ public class SubSelectFromDeclaration {
     this.declaration = new SelectFromDeclaration(context);
   }
 
-  public SubSelectFromDeclaration innerJoin(
+  @Override
+  public SelectContext get() {
+    return declaration.getContext();
+  }
+
+  public SubSelectFromDeclaration<ENTITY> innerJoin(
       EntityMetamodel<?> entityMetamodel, Consumer<JoinDeclaration> block) {
     declaration.innerJoin(entityMetamodel, block);
     return this;
   }
 
-  public SubSelectFromDeclaration leftJoin(
+  public SubSelectFromDeclaration<ENTITY> leftJoin(
       EntityMetamodel<?> entityMetamodel, Consumer<JoinDeclaration> block) {
     declaration.leftJoin(entityMetamodel, block);
     return this;
   }
 
-  public SubSelectFromDeclaration where(Consumer<WhereDeclaration> block) {
+  public SubSelectFromDeclaration<ENTITY> where(Consumer<WhereDeclaration> block) {
     declaration.where(block);
     return this;
   }
 
-  public SubSelectFromDeclaration orderBy(Consumer<OrderByNameDeclaration> block) {
+  public SubSelectFromDeclaration<ENTITY> groupBy(PropertyMetamodel<?>... propertyMetamodels) {
+    declaration.groupBy(propertyMetamodels);
+    return this;
+  }
+
+  public SubSelectFromDeclaration<ENTITY> having(Consumer<HavingDeclaration> block) {
+    declaration.having(block);
+    return this;
+  }
+
+  public SubSelectFromDeclaration<ENTITY> orderBy(Consumer<OrderByNameDeclaration> block) {
     declaration.orderBy(block);
     return this;
   }
 
-  public <PROPERTY> SubSelectContext<PROPERTY> select() {
+  public SubSelectContext<EntityMetamodel<ENTITY>> select() {
     SelectContext context = declaration.getContext();
-    return new SubSelectContext<>(context);
+    return () -> context;
   }
 
-  public <PROPERTY> SubSelectContext<PROPERTY> select(EntityMetamodel<?> entityMetamodel) {
+  public SubSelectContext<List<PropertyMetamodel<?>>> select(EntityMetamodel<?> entityMetamodel) {
     SelectContext context = declaration.getContext();
     context.projection = new Projection.PropertyMetamodels(entityMetamodel.allPropertyMetamodels());
-    return new SubSelectContext<>(context);
+    return () -> context;
   }
 
-  public <PROPERTY> SubSelectContext<PROPERTY> select(
-      PropertyMetamodel<PROPERTY> propertyMetamodel) {
+  public <PROPERTY> Single<PROPERTY> select(PropertyMetamodel<PROPERTY> propertyMetamodel) {
     SelectContext context = declaration.getContext();
     context.projection = new Projection.PropertyMetamodels(propertyMetamodel);
-    return new SubSelectContext<>(context);
+    return new Single<>(context, propertyMetamodel);
   }
 
-  public <PROPERTY1, PROPERTY2> SubSelectContext<Tuple2<PROPERTY1, PROPERTY2>> select(
-      PropertyMetamodel<PROPERTY1> first, PropertyMetamodel<PROPERTY2> second) {
+  public <PROPERTY1, PROPERTY2>
+      SubSelectContext<Tuple2<PropertyMetamodel<PROPERTY1>, PropertyMetamodel<PROPERTY2>>> select(
+          PropertyMetamodel<PROPERTY1> first, PropertyMetamodel<PROPERTY2> second) {
     SelectContext context = declaration.getContext();
     context.projection = new Projection.PropertyMetamodels(first, second);
-    return new SubSelectContext<>(context);
+    return () -> context;
   }
 
-  public SubSelectContext<List<?>> select(
+  public SubSelectContext<List<PropertyMetamodel<?>>> select(
       PropertyMetamodel<?> propertyMetamodel1,
       PropertyMetamodel<?> propertyMetamodel2,
       PropertyMetamodel<?>... propertyMetamodels) {
@@ -79,6 +94,6 @@ public class SubSelectFromDeclaration {
     list.add(propertyMetamodel2);
     list.addAll(Arrays.asList(propertyMetamodels));
     context.projection = new Projection.PropertyMetamodels(list);
-    return new SubSelectContext<>(context);
+    return () -> context;
   }
 }