+++ doma-processor/src/main/java/org/seasar/doma/internal/apt/cttype/BasicCtType.java
@@ -2,21 +2,29 @@ package org.seasar.doma.internal.apt.cttype;
 
 import static org.seasar.doma.internal.util.AssertionUtil.assertNotNull;
 
+import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.TypeMirror;
 import org.seasar.doma.internal.apt.Context;
 import org.seasar.doma.internal.apt.generator.Code;
+import org.seasar.doma.internal.jdbc.scalar.BasicScalarSuppliers;
+import org.seasar.doma.internal.jdbc.scalar.OptionalBasicScalarSuppliers;
+import org.seasar.doma.internal.util.Pair;
+import org.seasar.doma.internal.wrapper.WrapperSuppliers;
 
 public class BasicCtType extends AbstractCtType {
 
   private final TypeMirror boxedType;
 
+  private final TypeElement wrapperTypeElement;
+
   private final TypeMirror wrapperType;
 
-  BasicCtType(Context ctx, TypeMirror type, TypeMirror wrapperType) {
+  BasicCtType(Context ctx, TypeMirror type, Pair<TypeElement, TypeMirror> wrapperElementAndType) {
     super(ctx, type);
-    assertNotNull(wrapperType);
+    assertNotNull(wrapperElementAndType);
     this.boxedType = ctx.getMoreTypes().boxIfPrimitive(type);
-    this.wrapperType = wrapperType;
+    this.wrapperTypeElement = wrapperElementAndType.fst;
+    this.wrapperType = wrapperElementAndType.snd;
   }
 
   public TypeMirror getBoxedType() {
@@ -24,23 +32,27 @@ public class BasicCtType extends AbstractCtType {
   }
 
   public Code getWrapperSupplierCode() {
+    Class<?> clazz = WrapperSuppliers.class;
+    String methodName = "of" + wrapperTypeElement.getSimpleName().toString().replace("Wrapper", "");
     return new Code(
         p -> {
           if (isEnum()) {
-            p.print("() -> new %1$s(%2$s.class)", wrapperType, getQualifiedName());
+            p.print("%1$s.%2$s(%3$s.class)", clazz, methodName, getQualifiedName());
           } else {
-            p.print("%1$s::new", wrapperType);
+            p.print("%1$s.%2$s()", clazz, methodName);
           }
         });
   }
 
-  public Code getWrapperCode() {
+  public Code getScalarSupplierCode(boolean optional) {
+    Class<?> clazz = optional ? OptionalBasicScalarSuppliers.class : BasicScalarSuppliers.class;
+    String methodName = "of" + wrapperTypeElement.getSimpleName().toString().replace("Wrapper", "");
     return new Code(
         p -> {
           if (isEnum()) {
-            p.print("new %1$s(%2$s.class)", wrapperType, getQualifiedName());
+            p.print("%1$s.%2$s(%3$s.class)", clazz, methodName, getQualifiedName());
           } else {
-            p.print("new %1$s()", wrapperType);
+            p.print("%1$s.%2$s()", clazz, methodName);
           }
         });
   }