+++ test-criteria/src/test/java/example/NativeSqlUpdateTest.java
@@ -6,12 +6,16 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.seasar.doma.jdbc.criteria.expression.Expressions.add;
 import static org.seasar.doma.jdbc.criteria.expression.Expressions.concat;
 import static org.seasar.doma.jdbc.criteria.expression.Expressions.literal;
+import static org.seasar.doma.jdbc.criteria.expression.Expressions.max;
+import static org.seasar.doma.jdbc.criteria.expression.Expressions.select;
 
+import java.math.BigDecimal;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.seasar.doma.jdbc.Config;
 import org.seasar.doma.jdbc.SqlLogType;
 import org.seasar.doma.jdbc.criteria.NativeSql;
+import org.seasar.doma.jdbc.criteria.expression.SelectExpression;
 import org.seasar.doma.jdbc.criteria.statement.EmptyWhereClauseException;
 
 @ExtendWith(Env.class)
@@ -108,6 +112,20 @@ public class NativeSqlUpdateTest {
   }
 
   @Test
+  void expression_concat() {
+    Employee_ e = new Employee_();
+
+    int count =
+        nativeSql
+            .update(e)
+            .set(c -> c.value(e.employeeName, concat("[", concat(e.employeeName, "]"))))
+            .where(c -> c.eq(e.employeeId, 1))
+            .execute();
+
+    assertEquals(1, count);
+  }
+
+  @Test
   void expressions() {
     Employee_ e = new Employee_();
 
@@ -123,5 +141,34 @@ public class NativeSqlUpdateTest {
             .execute();
 
     assertEquals(1, count);
+  }
+
+  @Test
+  void expression_select() {
+    Employee_ e = new Employee_();
+    Employee_ e2 = new Employee_();
+    Department_ d = new Department_();
+
+    SelectExpression<Salary> subSelect =
+        select(
+            c ->
+                c.from(e2)
+                    .innerJoin(d, on -> on.eq(e2.departmentId, d.departmentId))
+                    .where(cc -> cc.eq(e.departmentId, d.departmentId))
+                    .groupBy(d.departmentId)
+                    .select(max(e2.salary)));
+
+    int count =
+        nativeSql
+            .update(e)
+            .set(c -> c.value(e.salary, subSelect))
+            .where(c -> c.eq(e.employeeId, 1))
+            .peek(System.out::println)
+            .execute();
+
+    assertEquals(1, count);
+
+    Salary salary = nativeSql.from(e).where(c -> c.eq(e.employeeId, 1)).select(e.salary).fetchOne();
+    assertEquals(0, new BigDecimal("3000").compareTo(salary.getValue()));
   }
 }